test: isolate durable usage history coverage
